Switch from boot2docker to docker-machine
The logic of the initialisation is fairly complex but is documented in
`?docker_machine_init`.  This works with the situation of >1 machine, but
falls back on the same behaviour as the previous `$(boot2docker shellinit)`
in the case of a single machine.

Because the docker-machine commands can be fairly slow, I've shifted them
to be conditional during .onLoad, based on the `harbor.autoconnect`
option (which can be set to `TRUE` or a machine name).